migration fix2

+++ migrations/000001_create_storage_table.up.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS storage (
     shortURL VARCHAR(255) PRIMARY KEY,
     originalURL VARCHAR(255) NOT NULL,
     count INTEGER NOT NULL,
-    uuid INTEGER NOT NULL,
+    uuid BIGINT NOT NULL,
     cookie VARCHAR(255) NOT NULL
 );
